nesting unnecessary - remove Inner component
